+++ huntflow_webhook_models/applicant.py
@@ -2,9 +2,9 @@ from typing import List
 
 from pydantic import BaseModel
 
-from models.base import BaseHuntflowWebhookRequest, WebhookMetaInfoBase
-from models.common_models.applicant import Applicant, ApplicantLog, ApplicantTag
-from models.consts import ApplicantWebhookActionType
+from huntflow_webhook_models.base import BaseHuntflowWebhookRequest, WebhookMetaInfoBase
+from huntflow_webhook_models.common_models.applicant import Applicant, ApplicantLog, ApplicantTag
+from huntflow_webhook_models.consts import ApplicantWebhookActionType
 
 
 class ApplicantEvent(BaseModel):
